refactor: implement load class of index (#5)
