+++ src/domain/index.ts
@@ -1,0 +1,6 @@
+import Config, { Icons } from './config'
+import DirList from './dirList'
+import RecentItem from './RecentItem'
+import ProjectRepository from './ProjectRepository'
+import ProjectQuickPick from './ProjectQuickPick'
+export { Config, Icons, DirList, RecentItem, ProjectRepository, ProjectQuickPick }
